requirements.txt
# Distillery — Knowledge Distillation con HuggingFace Trainer

# Core
transformers>=4.40.0
torch>=2.1.0
sentencepiece>=0.1.99

# Utilities
accelerate>=0.28.0    # necessario per Trainer multi-GPU / MPS
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
# AlphaGeometry / JAX (Necessari per l'estrazione dei logit)
jax>=0.4.0
jaxlib>=0.4.0
gin-config
flax
tqdm
